feat(config): add 'show toast' feature flags
Add the following flags for the user to enable/disable for controlling what notifications are shown.
  - bitburner.showPushSuccessNotification
  - bitburner.showFileWatcherEnabledNotification

Wrapper method for calling the vscode.window.* notification methods so we can have our branching logic of whether to show a notification or not in one place. We can pass in an optional arg and 'force show' the notification anyway, disregarding the users settings.

Added some 'TODO's for improvements in the future.